+++ CHANGELOG.md
@@ -6,8 +6,11 @@ Changelog for `exoplatform/exo:5.2.*_*` Docker image (older version : [5.1](./CH
 
 ## 5.2.0_0 [2019-04-04]
 
-- **Features**
+- **Upgrades**
   - upgrade to eXo Platform 5.2.0 GA
   - upgrade to Ubuntu 18.04 LTS (DOCKER-64)
   - upgrade to Libre Office 6.0 (DOCKER-65)
+
+- **Features**
   - remove support for Oracle (DOCKER-75)
+  - install PostgreSQL driver by default